requirements.txt
Django==3.2.5

django-treebeard==4.5.1
Pillow==8.3.1